+++ packages/vue/src/components/atoms/SfArrow/SfArrow.stories.js
@@ -1,8 +1,8 @@
 import { storiesOf } from "@storybook/vue";
 import {
   withKnobs,
+  boolean,
   optionsKnob as options,
-  object,
 } from "@storybook/addon-knobs";
 import { SfArrow, SfIcon } from "@storefront-ui/vue";
 storiesOf("Atoms|Arrow", module)
@@ -30,30 +30,13 @@ storiesOf("Atoms|Arrow", module)
           "CSS Modifiers"
         ),
       },
-      icon: {
-        default: object(
-          "icon",
-          {
-            icon: "arrow_left",
-            size: "1.5rem",
-          },
-          "Props"
-        ),
-      },
-      button: {
-        default: object(
-          "button",
-          {
-            disabled: false,
-          },
-          "Props"
-        ),
+      disabled: {
+        default: boolean("disabled", false, "Props"),
       },
     },
     template: `<SfArrow
       :class="customClass" 
-      :icon="icon"
-      :button="button"
+      :disabled="disabled"
       aria-label="Arrow label"
     />`,
   }))
@@ -78,28 +61,14 @@ storiesOf("Atoms|Arrow", module)
           "CSS Modifiers"
         ),
       },
-      icon: {
-        default: object(
-          "icon",
-          {
-            icon: "arrow_left",
-            size: "1.5rem",
-          },
-          "Props"
-        ),
-      },
-      button: {
-        default: object(
-          "button",
-          {
-            disabled: false,
-          },
-          "Props"
-        ),
+      disabled: {
+        default: boolean("disabled", false, "Props"),
       },
     },
     template: `<SfArrow
-      :class="customClass" aria-label="Arrow label">
+      :class="customClass"
+      :disabled="disabled"
+      aria-label="Arrow label">
       <SfIcon icon="chevron_left" class="sf-arrow__icon" size="14px" view-box="0 0 24 12"/>
     </SfArrow>`,
   }));