+++ packages/vue/src/components/atoms/SfBullets/SfBullets.vue
@@ -22,7 +22,7 @@
           :disabled="buttonActive.disabled"
           :link="buttonActive.link"
           aria-label="Current slide"
-          class="sf-button--pure sf-bullet sf-bullet--active"
+          class="sf-button--pure sf-bullet is-active"
         ></SfButton>
       </li>
     </slot>