+++ packages/vue/src/components/atoms/SfButton/SfButton.stories.js
@@ -21,6 +21,7 @@ storiesOf("Atoms|Button", module)
             "sf-button--pure": "sf-button--pure",
             "color-primary": "color-primary",
             "color-secondary": "color-secondary",
+            "color-light": "color-light",
             "color-warning": "color-warning",
             "color-danger": "color-danger",
             "color-info": "color-info",
@@ -39,7 +40,7 @@ storiesOf("Atoms|Button", module)
       },
       link: {
         default: text("link", "", "Props"),
-      }
+      },
     },
     components: { SfButton },
     template: `<SfButton
@@ -80,7 +81,7 @@ storiesOf("Atoms|Button", module)
       },
       link: {
         default: text("link", "https://www.storefrontui.io/", "Props"),
-      }
+      },
     },
     components: { SfButton },
     template: `<SfButton