+++ packages/vue/src/components/atoms/SfColor/SfColor.vue
@@ -1,7 +1,7 @@
 <template>
   <SfButton
     class="sf-button--pure sf-color"
-    :class="{ 'sf-color--active': selected }"
+    :class="{ 'is-active': selected }"
     :style="style"
     :disabled="button.disabled"
     :link="button.link"
@@ -14,9 +14,9 @@
       <slot name="badge" v-bind="{ selected, hasBadge }">
         <SfBadge
           v-if="selected && hasBadge"
-          class="sf-color__badge mobile-only"
+          class="sf-color__badge smartphone-only"
         >
-          <SfIcon aria-hidden="true" icon="check" size="8px" color="white" />
+          <SfIcon aria-hidden="true" icon="check" size="16px" color="white" />
         </SfBadge>
       </slot>
     </transition>