+++ packages/vue/src/components/atoms/SfInput/SfInput.vue
@@ -2,8 +2,8 @@
   <div
     class="sf-input"
     :class="{
-      'sf-input--has-text': !!value,
-      'sf-input--invalid': !valid,
+      'has-text': !!value,
+      invalid: !valid,
     }"
     :data-testid="name"
   >
@@ -45,7 +45,7 @@
           <SfIcon
             class="sf-input__password-icon"
             :class="{
-              'sf-input__password-icon--hidden': !isPasswordVisible,
+              hidden: !isPasswordVisible,
             }"
             :icon="icon.icon"
             :size="icon.size"