+++ packages/vue/src/components/atoms/SfQuantitySelector/SfQuantitySelector.stories.js
@@ -22,6 +22,8 @@ storiesOf("Atoms|QuantitySelector", module)
           {
             "sf-quantity-selector--secondary":
               "sf-quantity-selector--secondary",
+            "sf-quantity-selector--transparent":
+              "sf-quantity-selector--transparent",
           },
           "",
           { display: "multi-select" },