+++ packages/vue/src/components/molecules/SfColorPicker/SfColorPicker.vue
@@ -4,7 +4,7 @@
       <div
         v-if="!isOpen"
         key="color-picker-button"
-        class="sf-color-picker__button mobile-only"
+        class="sf-color-picker__button smartphone-only"
       >
         <!-- @slot Use this slot to replace open button. -->
         <slot name="open" v-bind="{ toggle }">
@@ -16,9 +16,13 @@
         </slot>
       </div>
       <div v-else key="color-picker-colors" class="sf-color-picker__colors">
+        <SfOverlay
+          :visible="isOpen"
+          class="sf-color-picker__colors__overlay smartphone-only"
+        />
         <!-- @slot Use this slot to replace label. -->
         <slot name="label" v-bind="{ label }">
-          <div v-if="label" class="sf-color-picker__label mobile-only">
+          <div v-if="label" class="sf-color-picker__label smartphone-only">
             {{ label }}
           </div>
         </slot>
@@ -28,7 +32,7 @@
         <slot name="close" v-bind="{ hasClose, toggle, isOpen }">
           <SfButton
             v-if="hasClose"
-            class="sf-button--text sf-color-picker__close mobile-only"
+            class="sf-button--text sf-color-picker__close smartphone-only"
             aria-label="Close button"
             :aria-pressed="!isOpen"
             @click="toggle"
@@ -43,9 +47,11 @@
 <script>
 import SfIcon from "../../atoms/SfIcon/SfIcon.vue";
 import SfButton from "../../atoms/SfButton/SfButton.vue";
+import SfOverlay from "../../atoms/SfOverlay/SfOverlay.vue";
+
 export default {
   name: "SfColorPicker",
-  components: { SfIcon, SfButton },
+  components: { SfIcon, SfButton, SfOverlay },
   props: {
     /**
      * ColorPicker is open