+++ packages/vue/src/components/molecules/SfPagination/SfPagination.vue
@@ -1,8 +1,8 @@
 <template>
   <nav class="sf-pagination">
     <!-- @slot Custom markup for previous page button -->
-    <slot name="prev" v-bind="{ isDisabled: !canGoPrev, prev: getPrev }">
-      <div class="sf-pagination__item sf-pagination__item--prev">
+    <slot name="prev" v-bind="{ isDisabled: !canGoPrev, go, prev: getPrev }">
+      <div class="sf-pagination__item prev">
         <component
           :is="componentIs"
           v-if="canGoPrev"
@@ -43,7 +43,7 @@
           class="sf-pagination__item"
           :class="{
             'sf-button--pure': !hasRouter && currentPage !== page,
-            'sf-pagination__item--current': currentPage === page,
+            current: currentPage === page,
           }"
           :link="hasRouter && currentPage !== page ? getLinkTo(page) : null"
           @click="!hasRouter && currentPage !== page ? go(page) : null"
@@ -71,8 +71,8 @@
       </slot>
     </template>
     <!-- @slot Custom markup for previous page button -->
-    <slot name="next" v-bind="{ isDisabled: !canGoNext, next: getNext }">
-      <div class="sf-pagination__item sf-pagination__item--next">
+    <slot name="next" v-bind="{ isDisabled: !canGoNext, go, next: getNext }">
+      <div class="sf-pagination__item next">
         <component
           :is="componentIs"
           v-if="canGoNext"
@@ -202,10 +202,8 @@ export default {
       this.$emit("click", page);
     },
     getLinkTo(page) {
-      return {
-        ...this.$route,
-        query: { ...this.$route.query, [this.pageParamName]: page },
-      };
+      const linkToPage = page.toString();
+      return linkToPage;
     },
   },
 };