+++ packages/vue/src/components/organisms/SfProductCard/SfProductCard.vue
@@ -68,7 +68,10 @@
           >
             <div class="sf-product-card__add-button--icons">
               <transition name="sf-pulse" mode="out-in">
-                <slot v-if="!isAddingToCart" name="add-to-cart-icon">
+                <slot
+                  v-if="!isAddingToCart && !isAddedToCart"
+                  name="add-to-cart-icon"
+                >
                   <SfIcon
                     key="add_to_cart"
                     icon="add_to_cart"
@@ -316,9 +319,7 @@ export default {
     },
     wishlistIconClasses() {
       const defaultClass = "sf-button--pure sf-product-card__wishlist-icon";
-      return `${defaultClass} ${
-        this.isOnWishlist ? "sf-product-card--on-wishlist" : ""
-      }`;
+      return `${defaultClass} ${this.isOnWishlist ? "on-wishlist" : ""}`;
     },
   },
   methods: {